+++ src/main/java/com/gmail/val59000mc/languages/Lang.java
@@ -148,7 +148,9 @@ public class Lang{
 	public static String TEAM_MESSAGE_COLOR_UNAVAILABLE;
 	public static String TEAM_MESSAGE_COLOR_CHANGED;
 	public static String TEAM_MESSAGE_NAME_CHANGED;
-	public static String TEAM_MESSAGE_NAME_CHANGED_ERROR;
+	public static String TEAM_MESSAGE_NAME_EMPTY;
+	public static String TEAM_MESSAGE_NAME_TOO_LONG;
+	public static String TEAM_MESSAGE_NAME_ILLEGAL_CHARACTERS;
 
 	public static String ITEMS_TEAM_LIST;
 	public static String ITEMS_SWORD;
@@ -376,7 +378,9 @@ public class Lang{
 		TEAM_MESSAGE_COLOR_UNAVAILABLE = getString(lang, "team.message.color-unavailable", "&cThis color is no longer available.");
 		TEAM_MESSAGE_COLOR_CHANGED = getString(lang, "team.message.color-changed", "&aYour team color was successfully changed.");
 		TEAM_MESSAGE_NAME_CHANGED = getString(lang, "team.message.name-changed", "&aYour team name was successfully changed.");
-		TEAM_MESSAGE_NAME_CHANGED_ERROR = getString(lang, "team.message.name-changed-error", "&cThat team name is not allowed!");
+		TEAM_MESSAGE_NAME_EMPTY = getString(lang, "team.message.name-empty", "&cThat team name is empty!");
+		TEAM_MESSAGE_NAME_TOO_LONG = getString(lang, "team.message.name-too-long", "&cThat team name is too long!");
+		TEAM_MESSAGE_NAME_ILLEGAL_CHARACTERS = getString(lang, "team.message.name-illegal-characters", "&cThat team name contains illegal characters!");
 
 		// Items
 		ITEMS_TEAM_LIST = getString(lang, "items.team-list", "&9Right click to view teams");